caixas de texto COMPLETAS

+++ Prognos/Model/CaixaTextoModel.swift
@@ -10,9 +10,9 @@ import SwiftUI
 struct CaixaTextoModel: Hashable {
     let placeholder: String
     var texto: String
-    let arredondamento: Int
+    let arredondamento: CGFloat
     
-    init(placeholder: String, texto: String, arredondamento: Int) {
+    init(placeholder: String, texto: String, arredondamento: CGFloat) {
         self.placeholder = placeholder
         self.texto = texto
         self.arredondamento = arredondamento
